refactor: float→int 坐标系统修复 + main.cpp 抽取 init_game() + 填充后殿/后花园场景
- rooms 键和 Prooms 从 float 改为 int，使用 ×10 整数坐标消除浮点精度问题
- where() 方向逻辑用 for 循环替代手写 s21 递增，N/S/W/E 坐标增量统一
- main.cpp: 抽取 init_game() 一次性初始化，添加 read_line_raw() 避免 raw 模式下 cin 冲突
- 注册随机事件 hell_whisper 和结局 abyss_gaze
- chapter3: 填充后殿(hd)/后花园(hh) 场景描述，去除 TODO 占位
- 防止 register_legacy_scenes() 重复注册

+++ include/story.h
@@ -7,10 +7,10 @@
 
 // ============ 全局变量声明 ============
 extern int random_val;
-extern std::map<float, std::string> rooms;
+extern std::map<int, std::string> rooms;
 extern std::map<int, char> alph;
 extern std::map<int, std::string> dire;
-extern float Prooms;
+extern int Prooms;
 
 extern bool chap1, chap2, chap3;
 extern std::string Chap1, Chap2, Chap3;
